+++ scripts/dwm-packages.sh
@@ -12,7 +12,7 @@ dwm_packages() {
 			libxcb xcb-util freetype2 fontconfig pkgconf
 		;;
 	arch:x11)
-		printf '%s\n' xorg-xinit xorg-xrandr xorg-xsetroot xorg-xset
+		printf '%s\n' xorg-xinit xorg-xrandr xorg-xsetroot xorg-xset xorg-xinput xorg-setxkbmap
 		;;
 	arch:x11-server)
 		printf '%s\n' xorg-server
@@ -73,9 +73,10 @@ dwm_packages() {
 	rhel:x11)
 		printf '%s\n' xorg-x11-server-Xorg xorg-x11-xinit
 		if [[ ${DISTRO_ID:-} == fedora ]]; then
-			printf '%s\n' xrandr xset xsetroot
+			printf '%s\n' xrandr xset xsetroot xinput setxkbmap
 		else
-			printf '%s\n' xorg-x11-server-utils
+			# Rocky/RHEL bundle xinput in xorg-x11-server-utils.
+			printf '%s\n' xorg-x11-server-utils setxkbmap
 		fi
 		;;
 	rhel:x11-server)
@@ -143,7 +144,7 @@ dwm_packages() {
 			libxcb1-dev libxcb-res0-dev libfontconfig-dev libfreetype-dev
 		;;
 	debian:x11)
-		printf '%s\n' xserver-xorg-core xinit x11-xserver-utils
+		printf '%s\n' xserver-xorg-core xinit x11-xserver-utils xinput x11-xkb-utils
 		;;
 	debian:x11-server)
 		: